Support for older versions of matplotlib.Figure (#199)
Setting the layout in Figure to constrained can be done in multiple ways. The version of matplotlib bundled with LCG 102 / current athena 23.0 is 3.3.something, and only supports the `constrained_layout=True` way, not `layout="constrained"`. Together with PR #198 this change makes the package fully compatible with latest numpy and matplotlib, and the versions currently pulled in with `asetup Athena,23.0`

Co-authored-by: Sam VS <sam.van.stroud@cern.ch>